Start providing facets in search results
This patch makes the searcher answer with facet couts. A simple query
for the top "vegan" recipe in the test data now yields something like
this:

    { "totalHits": 7,
      "recipes": [ {
          "recipeId": 293617,
          "name": "Vegan Noodle Salad with Peanut Sauce",
          "crawlUrl": "https://veggiedesserts.co.uk/vegan-noodle-salad-peanut-sauce-heck-thai-bites/" } ],
      "facets": [
          { "dimension": "Diet", "children": [ [ "vegan", 5 ] ] },
          { "dimension": "Keyword",
            "children": [
                [ "salt", 5 ],
                [ "oil", 4 ],
                [ "syrup", 4 ],
                [ "carrot", 3 ],
                [ "butter", 3 ],
                [ "sugar", 3 ],
                [ "maple syrup", 3 ],
                [ "seed", 3 ],
                [ "ginger", 2 ],
                [ "garlic", 2 ] ] } ]}

Which means that there are 7 documents that match the term "vegan",
out of those 5 have the label "vegan" (in the "Diet" facet) and
out of those, 3 have the "maple syrup" keyword.