Use SearchQuery.similarity instead of a flag
This patch makes changes how we do similarity searches: instead of
passing the text via `fulltext` and setting a `moreLikeThis` flag
in the json, we now simply use a `similarity` field.

We still don't allow fulltext AND similarity at the same time, but
now it's purely because of business logic, we can easily support
both at once now.