Add spring-boot dependency
Every ~100k recipes are ending up like ~100Mb in the index, so
I don't expect to need to shard anytime soon meaning that my
idea of using vert.x for its cute clustering functionalities
is kind of dead. This patch is the first in a series to move
to a simpler execution model.