Render responses in html
Turns out playing with more than one box gets expensive very
quickly. Who knew!

This patch introduces a dummy / endpoint, where the index page will
live; And a /search endpoint that acts as the previous one, but now
rendering an html result.

Using [rocker][1] as the template engine leads to some coupling as
I'll always either return something that is/contains
a RockerTemplate or is the output of it - I don't think it fits the
current model spring uses to decouple controllers (I can make it
slightly less coupled though - will do it in the future), but rocker
is such a neat little project that I'll try to keep it :-)

[1]: https://github.com/fizzed/rocker