WIP Metadata database
Flatbuffers + LMDB seem to work quite well together so far. And the
final database is ~700MB for 1+M recipes. This, along with an index
size of ~600MB leaves us with a very small ram requirement in runtime
and I can leave the java heap for all those frigging strings I can't
control.