Add descriptions to the lmdb+flatbuffers file
This obviously explodes the database size to ~2.2G - It is still
manageable, but I'm usure of how reasonable this is especially
since we pay a "heavy" price for strings when using flatbuffers
in java (more than we already pay garbage-wise :P) due to its
utf-8 encoding. So, maybe I'll keep it, maybe I'll look at
plain/manual serialization instead.