Use the new data model from cuisinier
This was a bit more involved than I would have liked, but the
good thing is that after all this time the separation between
UI and search is still there: notice no package under boot.*
had to be touched.

I might make this base into a multi-project maven setup so
that the search+db code is completely isolated, and that I
can iterate on document changes faster without building
silly profiles to skip web tests.

The new data model is a bit bigger (not accounting for an
increase in recipes quantity of about 3%):

New:

- 610M /index
- 2.8G /database

Old:

- 580M /index
- 2.4G /database