Revert "Stop using spring-boot-starter-parent as parent"
This reverts commit 0a72f9527905c4ca3af4e0577eaac6761e2894bf.

I've done it and decided to go back. These are the things I had to
do to make the multi-project setup happen:

1. Move the `co.caio.cerberus.boot` into a separate module (web)
2. Move the rest into a library module (core)
3. Create a parent pom with shared dependencies, move the
   non-shared dependencies around so that they belong to their
   respective module
4. Make the core library ship a test jar (as simple as activating
   the `test-jar` goal from maven-jar-plugin) so that both modules
   have access to the sample recipes and utilities
5. Change Util.java (test scope) to use .getResourceAsStream() so
   that we can static load it from #3's jar (*shrug* this is so
   poorly documented that I'm not really sure why .getResource()
   wasn't working.

(NOTE that this means there's a little bit of the strong coupling
I wanted to avoid by moving to multi-modules in the first place-
At least it is just for tests...)

After that was done I wanted to ship the `core` module separately
so that the `web` module would only evolve when I released a new
core version (same way as I'm interacting with `tablier`); This
led me to start fighting with `mvn release:*` since I doesn't
like SNAPSHOTS - so my solution would be to not have the `web`
module as part of the multi-module project, which meant that
the split was pretty much useless as I wouldn't be sharing things
from the parent POM unless I started being sillier...

Anyway, while doing all this and questioning my sanity I decided
that the best way would be to just split the repository and be
done with it. If I revisit this I'll probably:

1. Split the `boot` package into it's own repository; And
2. Create a separate project which just ships a parent pom
   in case I really want/need to share the pluginManagement
   section

But that's a story for another day, I've spent time with this for
no particular gain (besides understanding a bit more about maven,
so there's that). The result of this saga is in the
`dropped-multi-module` branch in case I ever need to look back.