Expose a &science= parameter to tweak the diet threshold
Now we're talking :-) I'm finally exposing the original motivator
for writing this whole thing: testing recipe diet classification.

I expect that with a high enough threshold we'll end up either with
recipes that truly belong to a diet group or that can easily be
tweaked to be part of it.