First incantation of SearchPolicy
This patch implements an interface that may be used to modify search
behaviour conditionally. This is intended for applying business
logic without having to expose an awkward/manual flow.

With this, the web UI will be able to:

* Block "MatchAllDocsQuery" without having to clean/munge the input
* Conditionally compute facets, reusing the (partial) search results
  in case getting the breakdowns is too expensive

I'm not particularly happy about exposing interfaces here, but
I can't think of a way to do this without going crazy on abstractions
for no particular gain.