Use SearchPolicy to inspect the parsed query only
All other parameters can be verified before executing the query,
so there's little use in forcing users of SearchPolicy to account
for all of that when inspecting the Query object.

This patch makes the QueryInterpreter aware of the policy and
in turn we now won't need to `query.rewrite(indexReader)` in
order to find a `MatchAllDocsQuery` or similar.