Drop QueryInterpreter: Move logic inside SearchImpl
It bothers me that I pass around a bunch of stuff just so that
I can create this QueryInterpreter that I don't have any intention
of providing alternative implementation for now.

The SearchPolicy addition in particular was what flagged that the
setup was quite awkward. Next step: Use inheritance to build a
searcher that handles policies instead of the current error
prone `searchPolicy != null`.