Make taxonomy facets configurable by the client
This commit makes the Indexer now accept an optional CategoryExtractor
that is responsible to extract categories and labels from any recipe.

What this effectively mean is that we support generic facets now and,
by default, we don't index any facet. CategoryExtractorTest has an
example on how this is all fitting together.

I'm not particularly happy about the state Searcher.Builder() is at
right now, but that's more cosmetic...