Persist IndexConfiguration to disk upon request
For some reason, FacetsConfig settings are not written and must be set
correctly otherwise counts will be wrong in some cases.

One alternative would be to pass the CategoryExtractor around, but
this means that I always need to know about how to extract facets even
if all I'm doing is querying; This patch takes an approach that I
consider better: we persist the necessary FacetsConfig setting to
the base directory we load from so that a Searcher will always do
the right thing.