WIP: Simple database to replace ChronicleMap
The main reason I'm using ChronicleMap is because I wanted an easy
Map interface with persistence and off-heap memory. Getting rid of
ChronicleMap reduces complexity, increases performance (though that
is negligible, db get()s are far from being the bottleneck) and
allows me to push Jdk12 forward.

This patch implements a working alternative which can be simplified
to a flat file with one recipe serialized (as a flatbuffer) after
another. In order to speed up loading I also write to an extra file
which contains the total number of recipes and recipe_id to offset
associations. The offset lookup table is kept in heap backed by
HPCC's primitive collections. Very little validation is done and
the code is totally susceptible to bad input attacks, but I have
full control of it, so :shrug:

A trivial benchmark such as:

```
public class MyBenchmark {

  @State(Scope.Benchmark)
  public static class MyState {

    RecipeMetadataDatabase chronicle;
    RecipeMetadataDatabase sdb;
    long[] ids =
        new long[] {289492, 707192, 1061982, 1708006, 1659287, 1653257, 901573, 1557621, 1639379};

    public MyState() {
      var cerberusPath = System.getProperty("cerberus");
      var sdbPath = System.getProperty("sdb");

      this.chronicle = ChronicleRecipeMetadataDatabase.open(Path.of(cerberusPath));
      this.sdb = new SimpleRecipeMetadataDatabase(Path.of(sdbPath));
    }
  }

  public MyBenchmark() {}

  private void check(RecipeMetadataDatabase db, long[] ids) {
    for (long id : ids) {
      var recipe = db.findById(id);
      assert recipe.isPresent();
      if (recipe.get().getRecipeId() != id) {
        throw new RuntimeException("oof!");
      }
    }
  }

  @Benchmark
  public void getChronicle(MyState state) {
    check(state.chronicle, state.ids);
  }

  @Benchmark
  public void getSdb(MyState state) {
    check(state.sdb, state.ids);
  }
}
```

Shows that at least things aren't broken. Yet.

> Benchmark                  Mode  Cnt        Score       Error  Units
> MyBenchmark.getChronicle  thrpt    5    62327.352 ±   214.437  ops/s
> MyBenchmark.getSdb        thrpt    5  2697423.234 ± 28008.573  ops/s