SearchQuery: Replace dietThreshold() with diet()
This patch changes the query model to only allow a single chosen diet
(and its corresponding threshold). This is being done mostly to simplify
usage: the only code I have that makes use of the feature is unpublished
and for reporting only - The web UI only uses one and I have no plans
of allowing more than one.